Separated constructors into separate files
